requirements.txt
Django==1.10.3
Flask==0.10.1
redis==2.10
behave==1.2.5
nose==1.3.7
pinocchio==0.4.2
noseOfYeti==1.6
expects==0.8.0
coverage==4.2
djangorestframework==3.5.3
django-filter==1.0.1
markdown==2.6.7
mysqlclient==1.3.6
